Introduce `NameMangler.API`
