+++ hs-bindgen/hs-bindgen.cabal
@@ -107,6 +107,7 @@ library
       HsBindgen.GenTests.Hs
       HsBindgen.GenTests.Internal
       HsBindgen.GenTests.Readme
+      HsBindgen.Hs.NameMangler.API
       HsBindgen.Hs.Translation
       HsBindgen.Imports
       HsBindgen.Orphans
